Update plugin-setting-wrapper.es
Dynamic reloading of plugins does not reflect the update in the Settings panel in case it goes from broken to fixed state - this PR solves this issue by assuming the problem has been fixed and allowing the error handlers to rethrow an error, if not. Otherwise it's forever stuck in hasError: true mode.